WL12XX: WLAN: OMAP2+: PANDA: Set MAC address in wl12xx's platform_data
This patch registers the Panda's WLAN unit bus path "mmc1:0001:2" with
the async platform_data api, so that device will be tagged with the
correct platform_data when it is probed.

This patch series adds a small helper in OMAP2+ id.c that generates a
stable per-board MAC address using CPU unique ID registers.

This patch uses the helper to set the MAC address of the wl12xx WLAN
device to a stable MAC address in the "locally administered" MAC space,
at the moment without this Panda WLAN comes up as 00:00:00:00:00:00
all the time.

Signed-off-by: Andy Green <andy.green@linaro.org>