ARM: OMAP5: GPIO: Implement workaround for Errata:i714
GPIO IRQ Not Generated After MPU Idle if IRQSTATUS
Bits Not Cleared:

DESCRIPTION:
After the GPIO is configured in smart-idle (or smart-idle with wake-up)
and the system goes into MPU inactive mode (idle), the GPIO does not
generate any IRQ again if any of the register bits of both interrupt line
raw status registers(GPIO_IRQSTATUS_RAW_0 or GPIO_IRQSTATUS_RAW_1) is set.
In the case of a GPIO configured in smart-idle wake-up mode
(GPIO_SYSCONFIG[4:3]=0x3), the wake-up associated to the GPIO IRQ event
will not even occur.

WORKAROUND:
The workaround is to clear GPIO_IRQSTATUS_y (all the bits) when the
corresponding interrupt line (most of the time the DSP, that is, 2nd
interrupt line) is not used, at each time GPIO_IRQSTATUS_x is cleared.

Signed-off-by: sricharan <r.sricharan@ti.com>