ARM: OMAP: add support for OMAP5 remoteproc cma sizes and start address modified
OMAP5 has to support a larger ipc shared memory region compared
to OMAP4. This region is reserved by CMA today using memblock
APIs. To account for the increase in size, the starting address
for the CMA resrvation needs to be different for OMAP5.

This patch adds the required support for OMAP5 and upholds
OMAP4 support too.

Modified for tilt kernel 3.4 integration

Signed-off-by: Subramaniam C.A <subramaniam.ca@ti.com>
Signed-off-by: Xavier Boudet <x-boudet@ti.com>