omap: rpmsg_resmgr: define and add opt clk definitions for iss
ISS needs an additional clock in order to enable the camera sensors
and become fully functional. This additional clock (iss_ctrlclk) is
different from the functional clock (iss_fck). The iss_fck is enabled
automatically through the device's pm_runtime_get_sync, but the opt
clock needs to be managed by the rpmsg resmgr driver. The name of this
this opt clock is published to the driver through the rpmsg resmgr
platform data.

Change-Id: I29f8aedd7cd8ca84be81b0e6c0d7cfcb7947a43d
Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>